Increasing Rack key space limit to handle longer keys
Since the introduction of the matrix questions, not only the payload for saving the forms, but also the key length for them is much larger since they include the parent question key, the x column key and y column key.
The combination of this was enough to exceed the default Rack key space limit that is in pace to prevent POST attacks.
We can't increase too much or we'll open the app for attacks, so we increased just enough to handle the current payloads that we have.