Reworking the validation on matrix questions as they are currently lagging a lot
Vlidations by default validate the entire question.
Some question types do have some special treatment, though.
Since the introduction of the matrix question, this behaviour has caused a significant lag when filling the form quickly.
The matrix question does not have the same HTML as other questions, which would make it behave correctly and validate only the input changed, and this caused the issue.
Sometimes to the point of completely freezing the page.

This commit fixes the issue by validating only the input that has last been changed, and only validating the entire question when it is expected to do so.
This commit also opens up a new parameter that can be later extended to other questions to behave this same way.