feat: Handle division by zero in growth calculation
This commit modifies the `FinancialTableSummary` module to handle division by zero in the growth calculation. Previously, if the previous value was zero, the calculation would result in an error. Now, if the previous value is zero, the growth percentage is set to "-" to indicate no growth. This prevents the division by zero error and ensures accurate growth calculations.

```

Signed-off-by: Louis Kirkham <louis.kirkham@bitzesty.com>