chore: Update brand name instructions in award year forms
This commit updates the instructions for the "brand_name" field in the award year forms. The new instructions clarify that the organization name should appear on the Royal award certificate and announcements. It also mentions that the award can be granted to a brand or trading name of the legal entity that has passed the KAE due diligence process, provided there is evidence of the legal entity using the trading name or owning the brand. The instructions also recommend not capitalizing all letters for better legibility on the award certificate.

Signed-off-by: Louis Kirkham <louis.kirkham@bitzesty.com>