chore: Update major_issues_overcome field in award year forms
This commit updates the help copy for the major_issues_overcome field in the award year forms. The wording has been changed to clarify that the field is specifically for major internal issues that have been overcome in recent years. It also provides an example of a significant issue that could affect the workforce, such as a major Health & Safety incident. The purpose of this update is to improve the clarity and understanding of the field for users.

Signed-off-by: Louis Kirkham <louis.kirkham@bitzesty.com>