WIP Replace pusher with anycable.
Each form_id/section should have a room, once a user connects to the room, that should then lock the form section for any other connection attempts to the same room, until the connection is disconnected.

It shouldnt be limited to collaboartor mode, as two browser tabs open will have an issue.